+++ automation/src/epilot/models/components/orconditionfordiff.py
@@ -6,7 +6,7 @@ from .anythingbutcondition import AnythingButCondition
 from .equalsignorecasecondition import EqualsIgnoreCaseCondition
 from .existscondition import ExistsCondition
 from .numericcondition import NumericCondition
-from .orcondition import OrCondition
+from .orcondition1 import OrCondition1
 from .prefixcondition import PrefixCondition
 from .suffixcondition import SuffixCondition
 from .wildcardcondition import WildcardCondition
@@ -18,6 +18,6 @@ from typing import Dict, List, Optional, Union
 @dataclass_json(undefined=Undefined.EXCLUDE)
 @dataclasses.dataclass
 class OrConditionForDiff:
-    dollar_or: Optional[List[Union[Union[OrCondition, Dict[str, List[Union[str, EqualsIgnoreCaseCondition, AnythingButCondition, NumericCondition, ExistsCondition, PrefixCondition, SuffixCondition, WildcardCondition]]]], Union[OrCondition, Dict[str, List[Union[str, EqualsIgnoreCaseCondition, AnythingButCondition, NumericCondition, ExistsCondition, PrefixCondition, SuffixCondition, WildcardCondition]]]], Union[OrCondition, Dict[str, List[Union[str, EqualsIgnoreCaseCondition, AnythingButCondition, NumericCondition, ExistsCondition, PrefixCondition, SuffixCondition, WildcardCondition]]]]]]] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('$or'), 'exclude': lambda f: f is None }})
+    dollar_or: Optional[List[Union[Union[OrCondition1, Dict[str, List[Union[str, EqualsIgnoreCaseCondition, AnythingButCondition, NumericCondition, ExistsCondition, PrefixCondition, SuffixCondition, WildcardCondition]]]], Union[OrCondition1, Dict[str, List[Union[str, EqualsIgnoreCaseCondition, AnythingButCondition, NumericCondition, ExistsCondition, PrefixCondition, SuffixCondition, WildcardCondition]]]], Union[OrCondition1, Dict[str, List[Union[str, EqualsIgnoreCaseCondition, AnythingButCondition, NumericCondition, ExistsCondition, PrefixCondition, SuffixCondition, WildcardCondition]]]]]]] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('$or'), 'exclude': lambda f: f is None }})